Made it compile on Mac. Tests fail due to multicast problem now.

+++ yogi-core/test/common.cc
@@ -216,7 +216,7 @@ TemporaryWorkdirGuard::~TemporaryWorkdirGuard() {
 
 CommandLine::CommandLine(std::initializer_list<std::string> args) {
   argc = static_cast<int>(args.size() + 1);
-  argv = new char*[argc];
+  argv = new char*[static_cast<std::size_t>(argc)];
 
   std::string exe = "executable-name";
   argv[0] = new char[exe.size() + 1];
